+++ CHANGELOG.md
@@ -5,11 +5,12 @@
 
 **Fixed**
 
+- **decidim-budgets**: Fixed bug where current projects's category wouldn't appear in form. [\#1666](https://github.com/decidim/decidim/pull/1666).
 - **decidim-core**: Fixed crash when running migrations & seeding DB in the same ruby process. [\#1649](https://github.com/decidim/decidim/pull/1649).
 - **decidim-core**: Fixed bug when blank values were used in a `datetime` field. [\#1661](https://github.com/decidim/decidim/pull/1661)
+- **decidim-core**: Fixed preview permissions for unpublished features. [\#1670](https://github.com/decidim/decidim/pull/1670).
 - **decidim-meetings**: Fixed bug where current meeting's category wouldn't appear in form. [\#1666](https://github.com/decidim/decidim/pull/1666).
 - **decidim-proposals**: Fixed bug where current proposal's category wouldn't appear in form. [\#1666](https://github.com/decidim/decidim/pull/1666).
-- **decidim-budgets**: Fixed bug where current projects's category wouldn't appear in form. [\#1666](https://github.com/decidim/decidim/pull/1666).
 - **decidim-results**: Fixed bug where current result's category wouldn't appear in form. [\#1666](https://github.com/decidim/decidim/pull/1666).
 
 ## [v0.4.3](https://github.com/decidim/decidim/tree/v0.4.3) (2017-07-25)